backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1